Separate main setup into dataswarm_worker_main.c
